+++ product-api/requirements.txt
@@ -1,0 +1,5 @@
+fastapi
+uvicorn
+psycopg2-binary
+SQLAlchemy
+python-dotenv
